LegacyNet/requirements.txt
PyQt5==5.15.1
pandas~=1.1.4
Pillow~=7.1.2
numpy~=1.18.5
matplotlib~=3.3.2
tensorflow~=2.3.1
tensorflow-object-detection-api~=0.1.1